created allReducers.js to combine reducers to export to store, moved tweetsReducer to a separate file
